+++ CHANGELOG.md
@@ -8,6 +8,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 # Unreleased
 
+# 4.10.1 (2022.06.22)
+
+## Changed
+
+- Upgraded dependencies to resolve security vulnerabilities. [#2075](https://github.com/stoplightio/prism/pull/2075) [#2076](https://github.com/stoplightio/prism/pull/2076)
+
+# 4.10.0 (2022.06.07)
+
+## Added
+
+- Added support for ranges of response status codes. [#2065](https://github.com/stoplightio/prism/pull/2065)
+
 # 4.9.3 (2022.05.19)
 
 ## Changed